修改各格式、migrations的位置以符合规范，并将jobplus/app.py合并到jobplus/__init.py
