+++ jobplus/handlers/__init__.py
@@ -1,5 +1,5 @@
 from .front import front
 from .admin import admin
 from .company import company
-from  .job import  job
-from  .user import user
+from .job import job
+from .user import user
