+++ jobplus/handlers/admin.py
@@ -1,4 +1,4 @@
-from flask import (Blueprint)
+from flask import Blueprint
 
 admin = Blueprint("admin", __name__, url_prefix='/admin')
