+++ jobplus/handlers/company.py
@@ -1,4 +1,4 @@
-from flask import (Blueprint)
+from flask import Blueprint
 
 company = Blueprint("company", __name__, url_prefix='/company')
