+++ jobplus/handlers/job.py
@@ -1,4 +1,4 @@
-from flask import (Blueprint)
+from flask import Blueprint
 
 job = Blueprint("job", __name__, url_prefix='/job')
