+++ jobplus/handlers/user.py
@@ -1,4 +1,4 @@
-from flask import (Blueprint)
+from flask import Blueprint
 
 user = Blueprint("user", __name__, url_prefix='/user')
