+++ manage.py
@@ -1,5 +1,5 @@
 import os
-from jobplus.app import create_app
+from jobplus import create_app
 from dotenv import load_dotenv
 
 dotenv_path = os.path.join(os.path.dirname(__file__), '.env')
@@ -7,12 +7,14 @@ if os.path.exists(dotenv_path):
     load_dotenv(dotenv_path)
 app = create_app('development')
 
+
 @app.cli.command()
 def test():
     """Run the unit tests."""
     import unittest
     tests = unittest.TestLoader().discover('tests')
     unittest.TextTestRunner(verbosity=2).run(tests)
+
 
 if __name__ == '__main__':
     app.run()
