+++ requirements.txt
@@ -3,22 +3,29 @@ blinker==1.4
 click==6.7
 dominate==2.3.1
 Flask==0.12.2
-Flask-Login==0.4.0
-Flask-Mail==0.9.1
 Flask-Bootstrap==3.3.7.1
+Flask-Login==0.4.0
 Flask-Migrate==2.0.4
 Flask-Moment==0.5.1
 Flask-SQLAlchemy==2.2
+Flask-Testing==0.7.1
 Flask-WTF==0.14.2
 itsdangerous==0.24
 Jinja2==2.9.6
 Mako==1.0.7
 MarkupSafe==1.0
+mysqlclient==1.3.14
+pycparser==2.19
+pymongo==3.7.2
+PyMySQL==0.9.2
 python-dateutil==2.6.1
-python-editor==1.0.3
 python-dotenv==0.10.1
+python-editor==1.0.3
+selenium==3.141.0
 six==1.10.0
 SQLAlchemy==1.1.11
+text-unidecode==1.2
+urllib3==1.24.1
 visitor==0.1.3
 Werkzeug==0.12.2
 WTForms==2.1
